refactor(tunings): adjust the response payload of tuning spec list
